Backend/requirements.txt
pandas==1.3.5
numpy==1.21.5
scikit-learn==1.0.2
lightgbm==3.3.2
xgboost==1.5.2
catboost==1.0.6
fastapi==0.70.0
uvicorn==0.15.0
sqlalchemy==1.4.29
psycopg2-binary==2.9.3
python-dotenv==0.19.2
joblib==1.1.0
optuna==2.10.0
shap==0.40.0
fastapi-jwt-auth==0.5.0
fastapi-limiter==0.1.5
redis==4.1.0
schedule==1.1.0
pytest==6.2.5
pydantic==1.8.2
gunicorn==20.1.0
azureml-sdk==1.41.0
azure-cosmos==4.3.0
azure-cli==2.33.1
confluent-kafka==1.8.2
pyspark==3.1.2
prometheus-client==0.14.1
opentelemetry-api==1.11.1
opentelemetry-sdk==1.11.1
opentelemetry-exporter-jaeger==1.11.1
networkx==2.8.4
python-louvain==0.16
vaderSentiment==3.3.2
pylint==2.14.5
scikit-learn==1.1.2
matplotlib==3.5.2
seaborn==0.11.2
schedule==1.1.0
tenacity==8.0.1
fastapi==0.68.0
uvicorn==0.15.0
grpcio==1.41.0
grpcio-tools==1.41.0
strawberry-graphql==0.96.0
protobuf==3.17.3
pydantic==1.8.2
